PoS Bug according to Maxim (5.2.X)
A possible bug was detected in the code for creating a stake block in the CreateCoinStake function (https://github.com/LUX-Core/lux/blob/02d3c429935ab0c0dd5f0342e2c64b99803fc4f5/src/stake.cpp#L1021)

The first argument passed to the CheckHash function is invalid. Instead of pindex-> pprev should be pIndex0-> pprev. As in the subsequent verification of the proof CheckProof (https://github.com/LUX-Core/lux/blob/02d3c429935ab0c0dd5f0342e2c64b99803fc4f5/src/stake.cpp#L736) the first argument is passed to the pointer to the previous block from the head.

To speed up the generation of stake blocks need to reduce the delay between calls to the coin selection function for minting and add a cycle of checking hashes in the time interval from the current moment plus 20-40 seconds to the future. This increases the probability of finding the desired hash which is less than the specified and will be approved by other participants. As a result, we are ahead of the rest in the minting speed without having large amounts on the balance.

The bug is present in the branches 5.2.X and 5.3.X.

Signed-off-by: 216k155 <32459326+216k155@users.noreply.github.com>